+++ en-US/resource.h
@@ -183,6 +183,8 @@
 #define IDS_CREATE_BUFFER               195
 #define IDS_NO_EVALUATION_SUPPORTED     196
 #define IDS_EVALUATION_FAILED           197
+#define IDS_TT_SORT_FIELD_SEPARATOR     205
+#define IDS_TT_SORT_FIELD_OPTIONS       206
 #define IDM_UNDO                        210
 #define IDM_DOAGAIN                     211
 #define IDM_ERASEBLOCK                  212
@@ -214,7 +216,7 @@
 // 
 #ifdef APSTUDIO_INVOKED
 #ifndef APSTUDIO_READONLY_SYMBOLS
-#define _APS_NEXT_RESOURCE_VALUE        200
+#define _APS_NEXT_RESOURCE_VALUE        205
 #define _APS_NEXT_COMMAND_VALUE         40010
 #define _APS_NEXT_CONTROL_VALUE         1011
 #define _APS_NEXT_SYMED_VALUE           101